+++ changelog.md
@@ -1,3 +1,9 @@
+
+## 2.1.0
+
+Standalone attachments added
+Change notification interface added
+
 ## 2.0.0
 
 Updated for Dart 1.0 release